data_processing/requirements.txt
paho-mqtt==1.6.1
APScheduler==3.9.1




